requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.6.20
cfgv==3.1.0
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==3.0.6
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-environ==0.4.5
django-import-export==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
filelock==3.0.12
identify==1.4.17
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
nodeenv==1.3.5
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.1.2
pre-commit==2.4.0
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
urllib3==1.25.9
uuid==1.30
virtualenv==20.0.21
xlrd==1.2.0
xlwt==1.3.0
